+++ docs/source/modules.rst
@@ -68,7 +68,7 @@ analyses/combinedwindwavesurgebuildingdamage
     :members:
 
 analyses/combinedwindwavesurgebuildingloss
-============================================
+==========================================
 ..  autoclass:: pyincore.analyses.combinedwindwavesurgebuildingloss.CombinedWindWaveSurgeBuildingLoss
     :members:
 
@@ -78,7 +78,7 @@ analyses/commercialbuildingrecovery
     :members:
 
 analyses/core_cge_ml
-===================================
+====================
 ..  autoclass:: core_cge_ml.corecgeml.CoreCGEML
     :members:
 
@@ -100,7 +100,7 @@ analyses/epfdamage
     :members:
 
 analyses/epfrepaircost
-==================
+======================
 ..  autoclass:: epfrepaircost.epfrepaircost.EpfRepairCost
     :members:
 
@@ -117,7 +117,7 @@ analyses/example
     :members:
 
 analyses/galvestoncge
-==================
+=====================
 ..  autoclass:: galvestoncge.galvestoncge.GalvestonCGEModel
     :members:
 ..  autoclass:: galvestoncge.equationlib.VarContainer
@@ -140,7 +140,7 @@ analyses/galvestoncge
     :members:
 
 analyses/gasfacilitydamage
-============================
+==========================
 ..  autoclass:: gasfacilitydamage.gasfacilitydamage.GasFacilityDamage
     :members:
 ..  autoclass:: gasfacilitydamage.gfutil.GfUtil
@@ -162,14 +162,14 @@ analyses/housingunitallocation
     :members:
 
 analyses/housingvaluationrecovery
-========================
+=================================
 ..  autoclass:: housingvaluationrecovery.housingvaluationrecovery.HousingValuationRecovery
     :members:
 ..  autoclass:: housingvaluationrecovery.housingvaluationrecoveryutil.HousingValuationRecoveryUtil
     :members:
 
 analyses/indp
-==============================
+=============
 ..  autoclass:: indp.indp.INDP
     :members:
 ..  autoclass:: indp.dislocationutils.DislocationUtil
@@ -215,7 +215,7 @@ analyses/joplincge
     :members:
 
 analyses/joplinempiricalbuildingrestoration
-===================================
+===========================================
 ..  autoclass:: joplinempiricalbuildingrestoration.joplinempiricalbuildingrestoration.JoplinEmpiricalBuildingRestoration
     :members:
 ..  autoclass:: joplinempiricalbuildingrestoration.joplinempirrestor_util.JoplinEmpirRestorUtil
@@ -227,13 +227,18 @@ analyses/meandamage
     :members:
 
 analyses/mlenabledcgeslc
-===================
+========================
 ..  autoclass:: mlenabledcgeslc.mlcgeslc.MlEnabledCgeSlc
     :members:
 
 analyses/montecarlofailureprobability
 =====================================
-..  autoclass:: montecarlofailureprobability.montecarlofailureprobability.MonteCarloFailureProbability
+.. deprecated:: 1.19.0
+   This class will be deprecated soon. Use :class:`montecarlolimitstateprobability.MonteCarloLimitStateProbability` instead.
+
+analyses/montecarlolimitstateprobability
+========================================
+..  autoclass:: montecarlolimitstateprobability.montecarlolimitstateprobability.MonteCarloLimitStateProbability
     :members:
 
 analyses/multiobjectiveretrofitoptimization
@@ -268,12 +273,12 @@ analyses/pipelinedamagerepairrate
     :members:
 
 analyses/pipelinefunctionality
-=========================
+==============================
 ..  autoclass:: pipelinefunctionality.pipelinefunctionality.PipelineFunctionality
     :members:
 
 analyses/pipelinerepaircost
-============================
+===========================
 ..  autoclass:: pipelinerepaircost.pipelinerepaircost.PipelineRepairCost
     :members:
 
@@ -315,7 +320,7 @@ analyses/seasidecge
     :members:
 
 analyses/saltlakecge
-==================
+====================
 ..  autoclass:: saltlakecge.saltlakecge.SaltLakeCGEModel
     :members:
 ..  autoclass:: saltlakecge.equationlib.VarContainer
@@ -345,7 +350,7 @@ analyses/socialvulnerability
     This class will be deprecated soon. Use :class:`socialvulnerabilityscore.SocialVulnerabilityScore` instead.
 
 analyses/socialvulnerabilityscore
-============================
+=================================
 ..  autoclass:: socialvulnerabilityscore.socialvulnerabilityscore.SocialVulnerabilityScore
     :members:
 
@@ -378,7 +383,7 @@ analyses/transportationrecovery
     This class will be deprecated soon. Use :class:`trafficflowrecovery.TrafficFlowRecovery` instead.
 
 analyses/trafficflowrecovery
-===============================
+============================
 ..  autoclass:: trafficflowrecovery.trafficflowrecovery.TrafficFlowRecovery
     :members:
 ..  autoclass:: trafficflowrecovery.trafficflowrecoveryutil.TrafficFlowRecoveryUtil
@@ -404,7 +409,7 @@ analyses/waterfacilitydamage
     :members:
 
 analyses/waterfacilityrepaircost
-=================================
+================================
 ..  autoclass:: waterfacilityrepaircost.waterfacilityrepaircost.WaterFacilityRepairCost
     :members:
 
@@ -414,7 +419,7 @@ analyses/waterfacilityrestoration
     :members:
 
 analyses/wfnfunctionality
-=================================
+=========================
 ..  autoclass:: wfnfunctionality.wfnfunctionality.WfnFunctionality
     :members:
 
@@ -422,7 +427,7 @@ models
 ^^^^^^
 
 models/hazard/earthquake
-===================
+========================
 ..  autoclass:: models.earthquake.Earthquake
     :members:
 
@@ -432,7 +437,7 @@ models/hazard/flood
     :members:
 
 models/hazard/hazard
-===================
+====================
 ..  autoclass:: models.hazard.Hazard
     :members:
 
@@ -487,7 +492,7 @@ models/mappingset
     :members:
 
 models/networkdataset
-=================
+=====================
 ..  autoclass:: models.networkdataset.NetworkDataset
     :members:
 
@@ -502,7 +507,7 @@ models/restorationcurveset
     :members:
 
 models/units
-==========================
+============
 ..  autoclass:: models.units.Units
     :members:
 
@@ -515,7 +520,7 @@ utils/analysisutil
     :members:
 
 utils/cge_ml_file_util
-==================
+======================
 ..  autoclass:: utils.cge_ml_file_util.CGEMLFileUtil
     :members:
